requirements.txt
requests>=2.31
python-dotenv>=1.0
chromadb>=0.4
pandas>=2.0